+++ Application/Ringtoets/test/Application.Ringtoets.Storage.Test/Read/Piping/SurfaceLineEntityReadExtensionsTest.cs
@@ -248,27 +248,27 @@ namespace Application.Ringtoets.Storage.Test.Read.Piping
                 {
                     new CharacteristicPointEntity
                     {
-                        X = x, Y = y, Z = z, Type = (short) CharacteristicPointType.BottomDitchDikeSide
+                        X = x, Y = y, Z = z, Type = (byte) CharacteristicPointType.BottomDitchDikeSide
                     },
                     new CharacteristicPointEntity
                     {
-                        X = x, Y = y, Z = z, Type = (short) CharacteristicPointType.BottomDitchPolderSide
+                        X = x, Y = y, Z = z, Type = (byte) CharacteristicPointType.BottomDitchPolderSide
                     },
                     new CharacteristicPointEntity
                     {
-                        X = x, Y = y, Z = z, Type = (short) CharacteristicPointType.DikeToeAtPolder
+                        X = x, Y = y, Z = z, Type = (byte) CharacteristicPointType.DikeToeAtPolder
                     },
                     new CharacteristicPointEntity
                     {
-                        X = x, Y = y, Z = z, Type = (short) CharacteristicPointType.DikeToeAtRiver
+                        X = x, Y = y, Z = z, Type = (byte) CharacteristicPointType.DikeToeAtRiver
                     },
                     new CharacteristicPointEntity
                     {
-                        X = x, Y = y, Z = z, Type = (short) CharacteristicPointType.DitchDikeSide
+                        X = x, Y = y, Z = z, Type = (byte) CharacteristicPointType.DitchDikeSide
                     },
                     new CharacteristicPointEntity
                     {
-                        X = x, Y = y, Z = z, Type = (short) CharacteristicPointType.DitchPolderSide
+                        X = x, Y = y, Z = z, Type = (byte) CharacteristicPointType.DitchPolderSide
                     }
                 }
             };
@@ -316,7 +316,7 @@ namespace Application.Ringtoets.Storage.Test.Read.Piping
         {
             return new CharacteristicPointEntity
             {
-                Type = (short) type,
+                Type = (byte)type,
                 X = point.X, Y = point.Y, Z = point.Z
             };
         }
